LPS-55581 Moves EditUserGroupAction - no logic changes
