+++ portal-impl/src/com/liferay/portlet/usergroupsadmin/UserGroupsAdminPortlet.java
@@ -15,9 +15,147 @@
 package com.liferay.portlet.usergroupsadmin;
 
 /**
+ * @author Charles May
  * @author Drew Brokke
  */
 
 public class UserGroupsAdminPortlet extends MVCPortlet {
+
+	@Override
+	public void processAction(
+			ActionMapping actionMapping, ActionForm actionForm,
+			PortletConfig portletConfig, ActionRequest actionRequest,
+			ActionResponse actionResponse)
+		throws Exception {
+
+		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
+
+		try {
+			if (cmd.equals(Constants.ADD) || cmd.equals(Constants.UPDATE)) {
+				updateUserGroup(actionRequest);
+			}
+			else if (cmd.equals(Constants.DELETE)) {
+				deleteUserGroups(actionRequest);
+			}
+
+			sendRedirect(actionRequest, actionResponse);
+		}
+		catch (Exception e) {
+			if (e instanceof PrincipalException) {
+				SessionErrors.add(actionRequest, e.getClass());
+
+				setForward(actionRequest, "portlet.user_groups_admin.error");
+			}
+			else if (e instanceof DuplicateUserGroupException ||
+					 e instanceof NoSuchUserGroupException ||
+					 e instanceof RequiredUserGroupException ||
+					 e instanceof UserGroupNameException) {
+
+				SessionErrors.add(actionRequest, e.getClass());
+
+				if (cmd.equals(Constants.DELETE)) {
+					String redirect = PortalUtil.escapeRedirect(
+						ParamUtil.getString(actionRequest, "redirect"));
+
+					if (Validator.isNotNull(redirect)) {
+						actionResponse.sendRedirect(redirect);
+					}
+				}
+			}
+			else {
+				throw e;
+			}
+		}
+	}
+
+	@Override
+	public ActionForward render(
+			ActionMapping actionMapping, ActionForm actionForm,
+			PortletConfig portletConfig, RenderRequest renderRequest,
+			RenderResponse renderResponse)
+		throws Exception {
+
+		try {
+			ActionUtil.getUserGroup(renderRequest);
+		}
+		catch (Exception e) {
+			if (e instanceof NoSuchUserGroupException ||
+				e instanceof PrincipalException) {
+
+				SessionErrors.add(renderRequest, e.getClass());
+
+				return actionMapping.findForward(
+					"portlet.user_groups_admin.error");
+			}
+			else {
+				throw e;
+			}
+		}
+
+		return actionMapping.findForward(
+			getForward(
+				renderRequest, "portlet.user_groups_admin.edit_user_group"));
+	}
+
+	protected void deleteUserGroups(ActionRequest actionRequest)
+		throws Exception {
+
+		long[] deleteUserGroupIds = StringUtil.split(
+			ParamUtil.getString(actionRequest, "deleteUserGroupIds"), 0L);
+
+		for (long deleteUserGroupId : deleteUserGroupIds) {
+			UserGroupServiceUtil.deleteUserGroup(deleteUserGroupId);
+		}
+	}
+
+	protected void updateUserGroup(ActionRequest actionRequest)
+		throws Exception {
+
+		long userGroupId = ParamUtil.getLong(actionRequest, "userGroupId");
+
+		String name = ParamUtil.getString(actionRequest, "name");
+		String description = ParamUtil.getString(actionRequest, "description");
+
+		ServiceContext serviceContext = ServiceContextFactory.getInstance(
+			UserGroup.class.getName(), actionRequest);
+
+		UserGroup userGroup = null;
+
+		if (userGroupId <= 0) {
+
+			// Add user group
+
+			userGroup = UserGroupServiceUtil.addUserGroup(
+				name, description, serviceContext);
+		}
+		else {
+
+			// Update user group
+
+			userGroup = UserGroupServiceUtil.updateUserGroup(
+				userGroupId, name, description, serviceContext);
+		}
+
+		// Layout set prototypes
+
+		long publicLayoutSetPrototypeId = ParamUtil.getLong(
+			actionRequest, "publicLayoutSetPrototypeId");
+		long privateLayoutSetPrototypeId = ParamUtil.getLong(
+			actionRequest, "privateLayoutSetPrototypeId");
+		boolean publicLayoutSetPrototypeLinkEnabled = ParamUtil.getBoolean(
+			actionRequest, "publicLayoutSetPrototypeLinkEnabled");
+		boolean privateLayoutSetPrototypeLinkEnabled = ParamUtil.getBoolean(
+			actionRequest, "privateLayoutSetPrototypeLinkEnabled");
+
+		if ((privateLayoutSetPrototypeId > 0) ||
+			(publicLayoutSetPrototypeId > 0)) {
+
+			SitesUtil.updateLayoutSetPrototypesLinks(
+				userGroup.getGroup(), publicLayoutSetPrototypeId,
+				privateLayoutSetPrototypeId,
+				publicLayoutSetPrototypeLinkEnabled,
+				privateLayoutSetPrototypeLinkEnabled);
+		}
+	}
 
 }
