--- portal-impl/src/com/liferay/portlet/usergroupsadmin/action/EditUserGroupAction.java
@@ -1,187 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portlet.usergroupsadmin.action;
-
-import com.liferay.portal.DuplicateUserGroupException;
-import com.liferay.portal.NoSuchUserGroupException;
-import com.liferay.portal.RequiredUserGroupException;
-import com.liferay.portal.UserGroupNameException;
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.model.UserGroup;
-import com.liferay.portal.security.auth.PrincipalException;
-import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.ServiceContextFactory;
-import com.liferay.portal.service.UserGroupServiceUtil;
-import com.liferay.portal.struts.PortletAction;
-import com.liferay.portal.util.PortalUtil;
-import com.liferay.portlet.sites.util.SitesUtil;
-
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-import javax.portlet.PortletConfig;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-
-import org.apache.struts.action.ActionForm;
-import org.apache.struts.action.ActionForward;
-import org.apache.struts.action.ActionMapping;
-
-/**
- * @author Charles May
- */
-public class EditUserGroupAction extends PortletAction {
-
-	@Override
-	public void processAction(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			ActionResponse actionResponse)
-		throws Exception {
-
-		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
-
-		try {
-			if (cmd.equals(Constants.ADD) || cmd.equals(Constants.UPDATE)) {
-				updateUserGroup(actionRequest);
-			}
-			else if (cmd.equals(Constants.DELETE)) {
-				deleteUserGroups(actionRequest);
-			}
-
-			sendRedirect(actionRequest, actionResponse);
-		}
-		catch (Exception e) {
-			if (e instanceof PrincipalException) {
-				SessionErrors.add(actionRequest, e.getClass());
-
-				setForward(actionRequest, "portlet.user_groups_admin.error");
-			}
-			else if (e instanceof DuplicateUserGroupException ||
-					 e instanceof NoSuchUserGroupException ||
-					 e instanceof RequiredUserGroupException ||
-					 e instanceof UserGroupNameException) {
-
-				SessionErrors.add(actionRequest, e.getClass());
-
-				if (cmd.equals(Constants.DELETE)) {
-					String redirect = PortalUtil.escapeRedirect(
-						ParamUtil.getString(actionRequest, "redirect"));
-
-					if (Validator.isNotNull(redirect)) {
-						actionResponse.sendRedirect(redirect);
-					}
-				}
-			}
-			else {
-				throw e;
-			}
-		}
-	}
-
-	@Override
-	public ActionForward render(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, RenderRequest renderRequest,
-			RenderResponse renderResponse)
-		throws Exception {
-
-		try {
-			ActionUtil.getUserGroup(renderRequest);
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchUserGroupException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(renderRequest, e.getClass());
-
-				return actionMapping.findForward(
-					"portlet.user_groups_admin.error");
-			}
-			else {
-				throw e;
-			}
-		}
-
-		return actionMapping.findForward(
-			getForward(
-				renderRequest, "portlet.user_groups_admin.edit_user_group"));
-	}
-
-	protected void deleteUserGroups(ActionRequest actionRequest)
-		throws Exception {
-
-		long[] deleteUserGroupIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "deleteUserGroupIds"), 0L);
-
-		for (long deleteUserGroupId : deleteUserGroupIds) {
-			UserGroupServiceUtil.deleteUserGroup(deleteUserGroupId);
-		}
-	}
-
-	protected void updateUserGroup(ActionRequest actionRequest)
-		throws Exception {
-
-		long userGroupId = ParamUtil.getLong(actionRequest, "userGroupId");
-
-		String name = ParamUtil.getString(actionRequest, "name");
-		String description = ParamUtil.getString(actionRequest, "description");
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			UserGroup.class.getName(), actionRequest);
-
-		UserGroup userGroup = null;
-
-		if (userGroupId <= 0) {
-
-			// Add user group
-
-			userGroup = UserGroupServiceUtil.addUserGroup(
-				name, description, serviceContext);
-		}
-		else {
-
-			// Update user group
-
-			userGroup = UserGroupServiceUtil.updateUserGroup(
-				userGroupId, name, description, serviceContext);
-		}
-
-		// Layout set prototypes
-
-		long publicLayoutSetPrototypeId = ParamUtil.getLong(
-			actionRequest, "publicLayoutSetPrototypeId");
-		long privateLayoutSetPrototypeId = ParamUtil.getLong(
-			actionRequest, "privateLayoutSetPrototypeId");
-		boolean publicLayoutSetPrototypeLinkEnabled = ParamUtil.getBoolean(
-			actionRequest, "publicLayoutSetPrototypeLinkEnabled");
-		boolean privateLayoutSetPrototypeLinkEnabled = ParamUtil.getBoolean(
-			actionRequest, "privateLayoutSetPrototypeLinkEnabled");
-
-		if ((privateLayoutSetPrototypeId > 0) ||
-			(publicLayoutSetPrototypeId > 0)) {
-
-			SitesUtil.updateLayoutSetPrototypesLinks(
-				userGroup.getGroup(), publicLayoutSetPrototypeId,
-				privateLayoutSetPrototypeId,
-				publicLayoutSetPrototypeLinkEnabled,
-				privateLayoutSetPrototypeLinkEnabled);
-		}
-	}
-
-}
